Beginner/file-organiser/requirements.txt
python-magic
rich